==> Workbook/3-Text Processing.ipynb <==
# Text Analysis through Natural Language Processing, Understanding and Classification

#Importing NLP packages
from wordcloud import WordCloud, STOPWORDS
from yellowbrick.text import FreqDistVisualizer
from sklearn.feature_extraction import stop_words
from textblob import TextBlob, Word
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
import re
import pandas as pd # to dataframes
import plotly.plotly as py
from plotly.offline import download_plotlyjs, init_notebook_mode, plot, iplot
import plotly.graph_objs as go

EDA=pd.read_excel('../CannaConnect/Dataset/EDA.xlsx')
EDA.head()

#Preprocessing the text in the Description, print the head of the description text
stopwords = set(STOPWORDS)
newStopWords = ['strain','effect', 'genetic', 'effects','flavor','dominant','known','cross'] #from the list above, we are keeping words relevant to the analysis
stopwords.update(newStopWords)
df_des=EDA.Description.astype(str)
df_lower_case_des=df_des.apply(lambda x: " ".join(x.lower() for x in x.split())) #lower case
df_special_characters_out_des=df_lower_case_des.map(lambda x: re.sub(r'\W+', ' ',x)) #remove special character
df_stop_words_out_des=df_special_characters_out_des.apply(lambda x: " ".join(x for x in x.split() if x not in stopwords)) #remove stopwords
df_lemma_des=df_stop_words_out_des.apply(lambda x: " ".join([Word(word).lemmatize() for word in x.split()]))  #lemmatize
print(df_lemma_des.head())

df_digits = df_lemma_des.apply(lambda x: len([x for x in x.split() if x.isdigit()]))
print(df_digits.head())
print(df_digits.describe())

df_numeric_out = df_lemma_des.str.replace('\\d+', '')
print(df_numeric_out.head())

#Step 1: Identification of Common words
#Step 2: I am not going to remove common words at this point
common_words = pd.Series(' '.join(df_stop_words_out_des).split()).value_counts()[:10]
print(common_words)

#Step 1: Identification of Rare words
#Step 2: I am not going to rare words at this point
rare_words = pd.Series(' '.join(df_stop_words_out_des).split()).value_counts()[-10:]
print(rare_words)

df_description_clean=df_numeric_out
df_description_clean.to_excel('../CannaConnect/Dataset/description_clean.xlsx')

#Visualization of the Frequent words
all_words = df_description_clean.str.split(expand=True).unstack().value_counts()
data = [go.Bar(
            x = all_words.index.values[1:50],
            y = all_words.values[1:50],
            marker= dict(colorscale='Jet',
                         color = all_words.values[2:100]
                        ),
            text='Word counts'
    )]

layout = go.Layout(
    title='Top 50 Word frequencies in the Dataset'
)

fig = go.Figure(data=data, layout=layout)

plot(fig, filename='basic-bar.html')
