==> Workbook/4-Sentiment Analysis.ipynb <==
# Importing package to be used

import pandas as pd # to dataframes
import matplotlib.pyplot as plt #to define plot parameters
import seaborn as sns #to graphical plots
import numpy as np #to math
from textblob import TextBlob, Word

EDA=pd.read_excel('../CannaConnect/Dataset/EDA.xlsx')
EDA.head()

df_description_clean=pd.read_excel('../CannaConnect/Dataset/description_clean.xlsx')
df_description_clean=df_description_clean.Description.astype(str)
df_description_clean.head()

# Sentiment analysis on Description, Effects and Flavor

#sentiment analysis on description with stopwords
EDA['Sentiment_Description']=EDA.Description.astype(str).apply(lambda x: TextBlob(x).sentiment.polarity)
print('Sentiment Description with stopwords', EDA.Sentiment_Description.head())

#sentiment analysis on description without stopwords and preprocessing
EDA['Sentiment_Description_clean']=df_description_clean.astype(str).apply(lambda x: TextBlob(x).sentiment.polarity)
print('Sentiment Description without stopwords ', EDA.Sentiment_Description_clean.head())

#sentiment analysis on Effects
EDA['Sentiment_Effects']=EDA.Effects.astype(str).apply(lambda x: TextBlob(x).sentiment.polarity)
print('Sentiment on Effects ', EDA.Sentiment_Effects.head())

#sentiment analysis on Flavors
EDA['Sentiment_Flavor']=EDA.Flavor.astype(str).apply(lambda x: TextBlob(x).sentiment.polarity)
print('Sentiment on Flavors ', EDA.Sentiment_Flavor.head())

# do analysis on polarity and on subjectivity

print('sentiment description count')
print('negative:', sum(i < 0 for i in EDA.Sentiment_Description))
print('positive:', sum(i > 0 for i in EDA.Sentiment_Description))
print('neutral:', sum(i == 0 for i in EDA.Sentiment_Description))
print('-------------')

print('sentiment Effects count')
print('negative:', sum(i < 0 for i in EDA.Sentiment_Effects))
print('positive:', sum(i > 0 for i in EDA.Sentiment_Effects))
print('neutral:', sum(i == 0 for i in EDA.Sentiment_Effects))
print('-------------')

print('sentiment Flavors count')
print('negative:', sum(i < 0 for i in EDA.Sentiment_Flavor))
print('positive:', sum(i > 0 for i in EDA.Sentiment_Flavor))
print('neutral:', sum(i == 0 for i in EDA.Sentiment_Flavor))
print('-------------')

EDA.head()

EDA.to_excel('../CannaConnect/Dataset/Sentiment_Analysis.xlsx')
