==> Workbook/6_2-Topic LDA.ipynb <==
# References:
#https://medium.com/mlreview/topic-modeling-with-scikit-learn-e80d33668730

import pandas as pd  
import numpy as np
documents=pd.read_excel("../CannaConnect/Dataset/description_clean.xlsx")
documents=documents.Description.astype(str)
documents.head()

from sklearn.feature_extraction.text import CountVectorizer

count_vect = CountVectorizer(max_df=0.95,min_df=5, stop_words='english')  
doc_term_matrix = count_vect.fit_transform(documents.values.astype('U'))
doc_term_matrix

from sklearn.decomposition import LatentDirichletAllocation

LDA = LatentDirichletAllocation(n_components=10, random_state=42)  
LDA.fit(doc_term_matrix)

import random

for i in range(10):  
    random_id = random.randint(0,len(count_vect.get_feature_names()))
    print(count_vect.get_feature_names()[random_id])

first_topic = LDA.components_[0]
top_topic_words = first_topic.argsort()[-10:]
for i in top_topic_words:  
    print(count_vect.get_feature_names()[i])

for i,topic in enumerate(LDA.components_):  
    print(f'Top 10 words for topic #{i}:')
    print([count_vect.get_feature_names()[i] for i in topic.argsort()[-10:]])
    print('\n')

topic_values = LDA.transform(doc_term_matrix)  
topic_values.shape

# NMF for Topic Modeling in Python

import pandas as pd  
import numpy as np
documents=pd.read_excel("../CannaConnect/Dataset/description_clean.xlsx")
documents=documents.Description.astype(str)
documents.head()

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vect = TfidfVectorizer(max_df=0.95, min_df=5, stop_words='english')  
doc_term_matrix = tfidf_vect.fit_transform(documents.values.astype('U'))

from sklearn.decomposition import NMF

nmf = NMF(n_components=10, random_state=42)  
nmf.fit(doc_term_matrix )

import random

for i in range(10):  
    random_id = random.randint(0,len(tfidf_vect.get_feature_names()))
    print(tfidf_vect.get_feature_names()[random_id])

first_topic = nmf.components_[0]  
top_topic_words = first_topic.argsort()[-10:]

for i in top_topic_words:  
    print(tfidf_vect.get_feature_names()[i])

for i,topic in enumerate(nmf.components_):  
    print(f'Top 10 words for topic #{i}:')
    print([tfidf_vect.get_feature_names()[i] for i in topic.argsort()[-10:]])
    print('\n')
