==> Workbook/8_0-Description to Type.ipynb <==
# From the Description can we predict the Type

import re
import numpy as np
import pandas as pd

from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import classification_report, confusion_matrix, mean_absolute_error
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer, TfidfTransformer
from sklearn.naive_bayes import GaussianNB, MultinomialNB
from sklearn.linear_model import LinearRegression, SGDClassifier, LogisticRegression


df_review=pd.read_excel("../CannaConnect/Dataset/description_clean.xlsx")
df=df_review.Description.astype(str)
df.head()

EDA=pd.read_excel('../CannaConnect/Dataset/EDA.xlsx')
EDA.head()

EDA=pd.DataFrame(EDA,columns=['Type'])
EDA.head()

des_type=pd.concat([df,EDA],axis=1)
des_type.shape

des_type=des_type.dropna()
des_type.shape

des_type.head()

# one hot encoding
lb_make = LabelEncoder()
des_type['Type'] = lb_make.fit_transform(des_type['Type'].astype(str))

X=des_type['Description']
y=des_type['Type']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

nb=Pipeline([('vect', CountVectorizer()),
            ('tfidf', TfidfTransformer()),
            ('clf', MultinomialNB()),
            ])
model=nb.fit(X_train, y_train)
print('Accuracy of NB classifier on training set: {:.2f}'.format(nb.score(X_train, y_train)))
print('Accuracy of NB classifier on test set: {:.2f}'.format(nb.score(X_test, y_test)))
y_pred=nb.predict(X_test)
print('MAE of NB classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, y_pred)))
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

sgd=Pipeline([('vect', CountVectorizer()),
            ('tfidf', TfidfTransformer()),
            ('clf', SGDClassifier()),
            ])
model=sgd.fit(X_train, y_train)
print('Accuracy of SGD classifier on training set: {:.2f}'.format(sgd.score(X_train, y_train)))
print('Accuracy of SGD classifier on test set: {:.2f}'.format(sgd.score(X_test, y_test)))
y_pred=sgd.predict(X_test)
print('MAE of SGD classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, y_pred)))
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

logreg=Pipeline([('vect', CountVectorizer()),
            ('tfidf', TfidfTransformer()),
            ('clf', LogisticRegression()),
            ])
model=logreg.fit(X_train, y_train)
print('Accuracy of LR classifier on training set: {:.2f}'.format(logreg.score(X_train, y_train)))
print('Accuracy of LR classifier on test set: {:.2f}'.format(logreg.score(X_test, y_test)))
y_pred=logreg.predict(X_test)
print('MAE of LR classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, y_pred)))
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
