==> Workbook/9_0-Sentiment to Rating.ipynb <==
# From the Sentiment analysis can we predict the Rating

import re
import numpy as np
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import classification_report, confusion_matrix, mean_absolute_error
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import GaussianNB, MultinomialNB
from sklearn.tree import DecisionTreeClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.linear_model import LinearRegression, SGDClassifier, LogisticRegression
from sklearn import utils, preprocessing

SA=pd.read_excel("../CannaConnect/Dataset/Sentiment_Analysis.xlsx")
SA=SA['Sentiment_Description']
SA.head()

Rating=pd.read_excel("../CannaConnect/Dataset/EDA.xlsx")
Rating=Rating['Rating']
Rating.head()

data=pd.concat([Rating,SA],axis=1)
data.shape

data=data.dropna()
data.shape

data.head()

data.dtypes

X=data[['Sentiment_Description']]
y=data[['Rating']].astype('int')
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
print(X_train.head())
print(X_train.shape)

#Getting rid of the the index values
X_train=X_train.values
y_train=y_train.values
X_test=X_test.values
y_test=y_test.values

#Logistic Regression
logreg=LogisticRegression().fit(X_train, y_train)
print('Accuracy of Logistic regression classifier on training set: {:.2f}'.format(logreg.score(X_train, y_train)))
print('Accuracy of Logistic regression classifier on test set: {:.2f}'.format(logreg.score(X_test, y_test)))
pred = logreg.predict(X_test)
print('MAE of Logistic regression classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, pred)))
print(confusion_matrix(y_test, pred))
print(classification_report(y_test, pred))
scores = cross_val_score(logreg, X, y, cv=10)
print("Cross-validated scores:", scores)

#Decision Tree
dtc = DecisionTreeClassifier().fit(X_train, y_train)
print('Accuracy of Decision Tree classifier on training set: {:.2f}'.format(dtc.score(X_train, y_train)))
print('Accuracy of Decision Tree classifier on test set: {:.2f}'.format(dtc.score(X_test, y_test)))
pred = dtc.predict(X_test)
print('MAE of Decision Tree classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, pred)))
print(confusion_matrix(y_test, pred))
print(classification_report(y_test, pred))
scores = cross_val_score(dtc, X, y, cv=10)
print("Cross-validated scores:", scores)

#KNeighbors
knn = KNeighborsClassifier().fit(X_train, y_train)
print('Accuracy of K-NN classifier on training set: {:.2f}'.format(knn.score(X_train, y_train)))
print('Accuracy of K-NN classifier on test set: {:.2f}'.format(knn.score(X_test, y_test)))
pred = knn.predict(X_test)
print('MAE of K-NN classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, pred)))
print(confusion_matrix(y_test, pred))
print(classification_report(y_test, pred))
scores = cross_val_score(knn, X, y, cv=10)
print("Cross-validated scores:", scores)

#Linear Discriminant Analysis
lda = LinearDiscriminantAnalysis().fit(X_train, y_train)
print('Accuracy of LDA classifier on training set: {:.2f}'.format(lda.score(X_train, y_train)))
print('Accuracy of LDA classifier on test set: {:.2f}'.format(lda.score(X_test, y_test)))
pred = lda.predict(X_test)
print('MAE of LDA classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, pred)))
print(confusion_matrix(y_test, pred))
print(classification_report(y_test, pred))
scores = cross_val_score(lda, X, y, cv=10)
print("Cross-validated scores:", scores)

#Gaussian Naive Bayes
gnb = GaussianNB().fit(X_train, y_train)
print('Accuracy of GNB classifier on training set: {:.2f}'.format(gnb.score(X_train, y_train)))
print('Accuracy of GNB classifier on test set: {:.2f}'.format(gnb.score(X_test, y_test)))
pred = gnb.predict(X_test)
print('MAE of GNB classifier on test set: {:.2f}'.format(mean_absolute_error(y_test, pred)))
print(confusion_matrix(y_test, pred))
print(classification_report(y_test, pred))
scores = cross_val_score(gnb, X, y, cv=10)
print("Cross-validated scores:", scores)
